move test utils to dedicated folder
