+++ src/test/kotlin/tech/relaycorp/relaynet/bindings/pdc/DetachedSignatureTypeTest.kt
@@ -8,9 +8,9 @@ import org.bouncycastle.asn1.DEROctetString
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import tech.relaycorp.relaynet.KeyPairSet
+import tech.relaycorp.relaynet.utils.KeyPairSet
 import tech.relaycorp.relaynet.OIDs
-import tech.relaycorp.relaynet.PDACertPath
+import tech.relaycorp.relaynet.utils.PDACertPath
 import tech.relaycorp.relaynet.crypto.SignedData
 import tech.relaycorp.relaynet.crypto.SignedDataException
 import tech.relaycorp.relaynet.wrappers.asn1.ASN1Utils
