+++ src/test/kotlin/tech/relaycorp/relaynet/messages/CargoCollectionAuthorizationTest.kt
@@ -1,7 +1,7 @@
 package tech.relaycorp.relaynet.messages
 
 import kotlin.test.Test
-import tech.relaycorp.relaynet.CERTIFICATE
+import tech.relaycorp.relaynet.utils.CERTIFICATE
 import tech.relaycorp.relaynet.ramf.RAMFSpecializationTestCase
 import tech.relaycorp.relaynet.wrappers.x509.Certificate
