+++ src/test/kotlin/tech/relaycorp/relaynet/messages/control/PrivateNodeRegistrationTest.kt
@@ -7,7 +7,7 @@ import org.bouncycastle.asn1.DERNull
 import org.bouncycastle.asn1.DEROctetString
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.assertThrows
-import tech.relaycorp.relaynet.PDACertPath
+import tech.relaycorp.relaynet.utils.PDACertPath
 import tech.relaycorp.relaynet.messages.InvalidMessageException
 import tech.relaycorp.relaynet.wrappers.asn1.ASN1Exception
 import tech.relaycorp.relaynet.wrappers.asn1.ASN1Utils
