+++ src/test/kotlin/tech/relaycorp/relaynet/messages/payloads/CargoCollectionRequestTest.kt
@@ -6,7 +6,7 @@ import org.bouncycastle.asn1.DEROctetString
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import tech.relaycorp.relaynet.CDACertPath
+import tech.relaycorp.relaynet.utils.CDACertPath
 import tech.relaycorp.relaynet.ramf.RAMFException
 import tech.relaycorp.relaynet.wrappers.asn1.ASN1Exception
 import tech.relaycorp.relaynet.wrappers.asn1.ASN1Utils
