+++ src/test/kotlin/tech/relaycorp/relaynet/messages/payloads/CargoMessageTest.kt
@@ -4,8 +4,8 @@ import kotlin.test.assertEquals
 import kotlin.test.assertNull
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
-import tech.relaycorp.relaynet.CERTIFICATE
-import tech.relaycorp.relaynet.KEY_PAIR
+import tech.relaycorp.relaynet.utils.CERTIFICATE
+import tech.relaycorp.relaynet.utils.KEY_PAIR
 import tech.relaycorp.relaynet.messages.Parcel
 import tech.relaycorp.relaynet.messages.ParcelCollectionAck
