+++ src/test/kotlin/tech/relaycorp/relaynet/messages/payloads/EncryptedPayloadTest.kt
@@ -6,8 +6,8 @@ import kotlin.test.assertEquals
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
-import tech.relaycorp.relaynet.CERTIFICATE
-import tech.relaycorp.relaynet.KEY_PAIR
+import tech.relaycorp.relaynet.utils.CERTIFICATE
+import tech.relaycorp.relaynet.utils.KEY_PAIR
 import tech.relaycorp.relaynet.SymmetricEncryption
 import tech.relaycorp.relaynet.wrappers.cms.EnvelopedData
 import tech.relaycorp.relaynet.wrappers.cms.PAYLOAD_SYMMETRIC_CIPHER_OIDS
