+++ src/test/kotlin/tech/relaycorp/relaynet/ramf/EncryptedRAMFMessageTest.kt
@@ -3,8 +3,8 @@ package tech.relaycorp.relaynet.ramf
 import kotlin.test.assertEquals
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
-import tech.relaycorp.relaynet.CERTIFICATE
-import tech.relaycorp.relaynet.issueStubCertificate
+import tech.relaycorp.relaynet.utils.CERTIFICATE
+import tech.relaycorp.relaynet.utils.issueStubCertificate
 import tech.relaycorp.relaynet.messages.payloads.StubEncryptedPayload
 import tech.relaycorp.relaynet.wrappers.generateRSAKeyPair
