+++ src/test/kotlin/tech/relaycorp/relaynet/ramf/RAMFMessageTest.kt
@@ -11,10 +11,10 @@ import kotlin.test.assertTrue
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.assertThrows
 import tech.relaycorp.relaynet.HashingAlgorithm
-import tech.relaycorp.relaynet.KeyPairSet
-import tech.relaycorp.relaynet.PDACertPath
-import tech.relaycorp.relaynet.assertDateIsAlmostNow
-import tech.relaycorp.relaynet.issueStubCertificate
+import tech.relaycorp.relaynet.utils.KeyPairSet
+import tech.relaycorp.relaynet.utils.PDACertPath
+import tech.relaycorp.relaynet.utils.assertDateIsAlmostNow
+import tech.relaycorp.relaynet.utils.issueStubCertificate
 import tech.relaycorp.relaynet.messages.InvalidMessageException
 import tech.relaycorp.relaynet.wrappers.cms.HASHING_ALGORITHM_OIDS
 import tech.relaycorp.relaynet.wrappers.cms.parseCmsSignedData
