+++ src/test/kotlin/tech/relaycorp/relaynet/ramf/RAMFSerializerTest.kt
@@ -23,11 +23,11 @@ import org.bouncycastle.asn1.DERGeneralizedTime
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import tech.relaycorp.relaynet.BER_DATETIME_FORMATTER
+import tech.relaycorp.relaynet.utils.BER_DATETIME_FORMATTER
 import tech.relaycorp.relaynet.HashingAlgorithm
 import tech.relaycorp.relaynet.crypto.SignedData
 import tech.relaycorp.relaynet.crypto.SignedDataException
-import tech.relaycorp.relaynet.issueStubCertificate
+import tech.relaycorp.relaynet.utils.issueStubCertificate
 import tech.relaycorp.relaynet.wrappers.asn1.ASN1Utils
 import tech.relaycorp.relaynet.wrappers.cms.HASHING_ALGORITHM_OIDS
 import tech.relaycorp.relaynet.wrappers.cms.parseCmsSignedData
