+++ src/test/kotlin/tech/relaycorp/relaynet/ramf/RAMFSpecializationTestCase.kt
@@ -4,8 +4,8 @@ import java.time.ZonedDateTime
 import kotlin.test.assertEquals
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
-import tech.relaycorp.relaynet.CERTIFICATE
-import tech.relaycorp.relaynet.issueStubCertificate
+import tech.relaycorp.relaynet.utils.CERTIFICATE
+import tech.relaycorp.relaynet.utils.issueStubCertificate
 import tech.relaycorp.relaynet.wrappers.generateRSAKeyPair
 import tech.relaycorp.relaynet.wrappers.x509.Certificate
