+++ src/test/kotlin/tech/relaycorp/relaynet/utils/DateUtils.kt
@@ -1,4 +1,4 @@
-package tech.relaycorp.relaynet
+package tech.relaycorp.relaynet.utils
 
 import java.time.ZoneId
 import java.time.ZonedDateTime
