+++ src/test/kotlin/tech/relaycorp/relaynet/utils/MockPrivateKeyStore.kt
@@ -1,4 +1,7 @@
-package tech.relaycorp.relaynet.keystores
+package tech.relaycorp.relaynet.utils
+
+import tech.relaycorp.relaynet.keystores.PrivateKeyData
+import tech.relaycorp.relaynet.keystores.PrivateKeyStore
 
 class MockPrivateKeyStore(
     private val savingException: Throwable? = null,
