+++ src/test/kotlin/tech/relaycorp/relaynet/utils/PDACertPath.kt
@@ -1,6 +1,9 @@
-package tech.relaycorp.relaynet
+package tech.relaycorp.relaynet.utils
 
 import java.time.ZonedDateTime
+import tech.relaycorp.relaynet.issueDeliveryAuthorization
+import tech.relaycorp.relaynet.issueEndpointCertificate
+import tech.relaycorp.relaynet.issueGatewayCertificate
 
 /**
  * Full certification path from a public gateway to a PDA.
